requirements.txt
PyMuPDF 
python-docx 
Pillow 
pytesseract
mimetypes
fitz